Refactor tests using router provider to fix change
tanstack router 117.1 fixed a bug/issue which broke this way of doing
tests due to it not properly checking load functions for the first
render. This broke our snapshot tests. So we implement a custom
render function that makes sure to wait for the router to load before
rendering with @testing-library/react.